requirements.txt
pynest-api==0.2.2    
bcrypt>=4.0.1
PyJWT>=2.8.0
passlib>=1.7.4
requests>=2.31.0
Faker>=19.2.0
redis>=5.0.0
yfinance>=0.2.27